#1
Just the base of the simple index for perrystowingweb.html with some js to make an example service call and some minimal css.  The index.php is so that heroku will automatically deploy it as a php app and that will point back to the specified perrystowingweb.html.

The ajax call will not work as it is now.  The server does not respond with the correct CORS settings.  Needs to be updated.